- Re-organise routers - Manage Shell branches for desktop navigation
