+++ app/lib/features/home/data/models/nav_item.dart
@@ -14,9 +14,11 @@ class BottomBarNavigationItem extends BottomNavigationBarItem {
 class SidebarNavigationItem extends NavigationRailDestination {
   final String? location;
   final bool pushToNavigate;
+  final bool isSpaceTab;
   const SidebarNavigationItem({
     this.location,
     this.pushToNavigate = false,
+    this.isSpaceTab = false,
     required Widget icon,
     required Widget label,
   }) : super(icon: icon, label: label);
